+++ flama/authentication/__init__.py
@@ -1,3 +1,3 @@
 from flama.authentication.components import *  # noqa
 from flama.authentication.jwt import *  # noqa
-from flama.authentication.middlewares import *  # noqa
+from flama.authentication.middleware import *  # noqa
